\name{close8graph}
\alias{close8graph}
\title{
(00) close the file open by open8graph
}
\description{
According to the global constant rbsb.mfi close the file open by
open8graph. Also if \code{!rbsb.mba} a pause is issued.
}
\usage{
close8graph(message=rbsb.msi)
}
\arguments{
\item{message}{ Message to display on the terminal.}
}
\value{
nothing but the actions indicated in the description field are
performed
}
\keyword{IO}
