\name{easyp3stickback}
\alias{easyp3stickback}
\title{
(f0) converse operation of easyp3cut
}
\description{
d must be a list with \$typ and \$blo components (see easyp3cut for
the details)
}
\usage{
easyp3stickback(d)
}
\arguments{
\item{d}{d\$typ for the type of components, d\$blo for the content of
each component.}
}
\details{
The braces used are defined into the matrix of constants
options()\$rbsb.cpt. Those used are precised in the local vector
quels.\cr
}
\value{
A character(1) of the reconstituted easyprograming
}
\keyword{utilities}
