\name{expr3func}
\alias{expr3func}
\title{
(f0) provides the generation function from gc and ft
}
\description{
From pieces of code usually generated by easyp2code, respectively for
generation and transformation, the generation function is produced.
}
\usage{
expr3func(gc, tc="")
}
\examples{
 expr3func("{rnorm(100)}", "{abs(Y)}");
}
\arguments{
\item{gc}{code for generation, character of length 1.}
\item{tc}{Code for transformation, character of length 1. When empty
transformation part is introduced.}
}
\details{
No check is made about the code validity.
}
\value{
resulting object function (to be introduced in the @rlks slot of the
bn's
}
\keyword{utilities}
