\name{form3etsil}
\alias{form3etsil}
\title{
(00) inverse function of form3liste or whatever
}
\description{
returns a vector character of the names from a character string
generated by form3liste. For the moment, sep cannot be an empty
string.\cr Of course, it is implicetly supposed that the inversion is
unambiguous.
}
\usage{
form3etsil(cara, none="-.-", OPA="{", CPA="}", opa="(", cpa=")", sep="+")
}
\examples{
 uu <- form3liste(letters[1:4]);
 form3etsil(uu);
}
\arguments{
\item{cara}{Character to be considered.}
\item{none}{ idem as in form3liste.}
\item{OPA}{ idem as in form3liste.}
\item{CPA}{ idem as in form3liste.}
\item{opa}{ idem as in form3liste.}
\item{cpa}{ idem as in form3liste.}
\item{sep}{ idem as in form3liste.}
}
\details{
The consistence of the different parenthesis and separators is
checked: an error is issued when they does not correspond.\cr
Important: the possible \\n added by form3liste is not taken into
account and must be removed before calling this function.
}
\value{
A character vector or nothing when imp is TRUE
}
\keyword{print}
