\name{rapport}
\alias{rapport}
\title{
(00) issues an error message when rebastaba fails
}
\description{
Ask the user to send the author a detailed report.
}
\usage{
rapport(origine)
}
\examples{
 # rapport("For the moment, we are doing the tests...");
}
\arguments{
\item{origine}{A character indicating where the difficulty occured.}
}
\value{
nothing
}
\keyword{error}
