\name{close8graph}
\alias{close8graph}
\title{
(f0) close the file open by open8graph
}
\description{
According to the global constant rbsb3k("mfi") close the file open by
open8graph. Also if \code{!rbsb3k("mba")} a pause is issued.
}
\usage{
close8graph(message=rbsb3k("msi"))
}
\arguments{
\item{message}{ Message to display on the terminal.}
}
\value{
nothing but the actions indicated in the description field are
performed
}
\keyword{IO}
