\name{close8text}
\alias{close8text}
\title{
(f0) pauses (and more) the program until an answer is given
}
\description{
Closing the output file rbsb3k("fou") according to rbsb3k("mfi"). A
pause allowing to stop the process is issued if rbsb3k("mba") is
false.
}
\usage{
close8text(message=rbsb3k("msi"))
}
\arguments{
\item{message}{ Message to display on the terminal.}
}
\value{
nothing but the actions indicated in the description field are
performed
}
\keyword{IO}
