\name{form3affiche}
\alias{form3affiche}
\title{
(00) displays with its name the object x
}
\description{
everything is in the title
}
\usage{
form3affiche(x, pau=FALSE, cat=FALSE, ...)
}
\examples{
 uu <- "azerty";
 form3affiche(uu);
}
\arguments{
\item{x}{The object to print.}
\item{pau}{ Must a pause be performed after the display?}
\item{cat}{ Must the printing be done with 'cat' instead of print?}
\item{\dots}{possible arguments for the print function.}
}
\value{
a print (or cat) is done and x is returned
}
\keyword{print}
